% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findWIKSIgroups.R
\name{findWISKIgroups}
\alias{findWISKIgroups}
\title{Finds all groups on the Wiski server}
\usage{
findWISKIgroups(site.url = "http://giws.usask.ca:8080/")
}
\arguments{
\item{site.url}{Optional. A character string containing the url of the WISKI web server. Defaults to the Global Institute for Water Security (GIWS) server: \code{'http://giws.usask.ca:8080/'}. As this package is intended for use by the GIWS hydrological community, it is usually unnecessary to specify the web server.}
}
\value{
If unsuccessful, returns \code{FALSE}. If successful, returns a dataframe containing \item{group_id}{ID number of the group} \item{group_name}{WISKI name of the group} \item{gorup_type}{Type of the group}
}
\description{
Finds all groups on the Wiski server
}
\examples{
groups <- findWISKIgroups()
}
\seealso{
\code{\link{findWISKIstations}}  \code{\link{findWISKItimeseries}}
}
\author{
Kevin Shook
}
