% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findWIKSIstations.R
\name{findWISKIstations}
\alias{findWISKIstations}
\title{Finds all stations on the Wiski server}
\usage{
findWISKIstations(stationName = "*",
  site.url = "http://giws.usask.ca:8080/")
}
\arguments{
\item{stationName}{Required. Name of station to search for. May contain wildcards. The defaul is \option{*}, which will search for all of the stations on the server.}

\item{site.url}{Optional. A character string containing the url of the WISKI web server. Defaults to the Global Institute for Water Security (GIWS) server: \code{'http://giws.usask.ca:8080/'}. As this package is intended for use by the GIWS hydrological community, it is usually unnecessary to specify the web server.}
}
\value{
If unsuccessful, returns \code{FALSE}. If successful, returns a dataframe containing \item{station_name}{Name of station, specified by \code{stationName}} \item{station_no}{WISKI number of station} \item{station_id}{WISKI station ID number} \item{station_latitude}{Latitude of station} \item{station_longitude}{Longitude of station}
}
\description{
Finds all stations on the Wiski server
}
\examples{
stations <- findWISKIstations()
}
\author{
Kevin Shook
}
\seealso{
\code{\link{findWISKItimeseries}} \code{\link{findWISKIgroups}}
}

