% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWISKImetadata.R
\name{getWISKImetadata}
\alias{getWISKImetadata}
\title{Gets metadata of a WISKI time series}
\usage{
getWISKImetadata(timeSeries, site.url = "http://giws.usask.ca:8080/")
}
\arguments{
\item{timeSeries}{Required. Character string containing the WISKI time series ID number, which is returned by \code{findWISKItimeseries}. Cannot contain wild card characters.}

\item{site.url}{Optional. A character string containing the url of the WISKI web server. Defaults to the Global Institute for Water Security (GIWS) server: \code{'http://giws.usask.ca:8080/'}. As this package is intended for use by the GIWS hydrological community, it is usually unnecessary to specify the web server.}
}
\value{
If unsucessful, returns \code{FALSE}. If successful, returns a dataframe containing the metadata of the specified time series.
}
\description{
Returns metadata of a specified time series.
}
\examples{
# Finds metadata for Fisera Ridge air temperatures (Original time series) 
FiseraTmetadata <- getWISKImetadata('9296042')
print(FiseraTmetadata)
}
\author{
Kevin Shook
}
\seealso{
\code{\link{getWISKIvalues}} \code{\link{findWISKItimeseries}}
}

