% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WISKIr-package.R
\docType{package}
\name{WISKIr-package}
\alias{WISKIr-package}
\title{Contains functions to extract data from a Kisters WISKI web server.}
\description{
Using the functions, data values can be downloaded directly from a WISKI web server. This allows you to automate data aquisition in your scripts. It is a good idea to use the function \code{\link{findWISKItimeseries}} first. It allows you to search by the name of a station, and gives you the \code{id} numbers for for all of its time series. When you have the code{id} number of a time series, you can obtain its metadata with the command \code{\link{getWISKImetadata}}. The actual values are downloaded using the command \code{\link{getWISKIvalues}}. The WISKI values can be converted to a \pkg{CRHMr} obs dataframe using the command \code{\link{WISKItoObs}}.
}
\references{
To cite \pkg{WISKIr} in publications, use the command \code{citation('WISKIr')} to get the current version of the citation.
}

