% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WISKItoObs.R
\name{WISKItoObs}
\alias{WISKItoObs}
\title{Converts a WISKI time series to an obs dataframe}
\usage{
WISKItoObs(WISKItimeseries, timezone = "", varName = "")
}
\arguments{
\item{WISKItimeseries}{Required. Dataframe containing the WISKI time series, as returned by getWISKIvalues.}

\item{timezone}{Required. The name of the timezone of the data as a character string. This should be the timezone of your data, but omitting daylight savings time. Note that the timezone code is specific to your OS. To avoid problems, you should use a timezone without daylight savings time. Under Linux, you can use \option{CST} and \option{MST} for Central Standard or Mountain Standard time, respectively. Under Windows or OSX, you can use \option{etc/GMT+6} or \option{etc/GMT+7} for Central Standard and Mountain Standard time. DO NOT use \option{America/Regina} as the time zone, as it includes historical changes between standard and daylight savings time.}

\item{varName}{Optional. Name for the variable. If not specified (the default) then a name will be guessed from the units. Note that many different variables have the same units, so the guess might not be very good!}
}
\description{
Converts the WISKI date/time to a CRHM datetime. If possible, it will also convert the variable name to a CRHM variable name. Note that this may be incorrect, as it is attempting to infer the variable from its units.
}
\examples{
FiseraTvalues <- getWISKIvalues('9296042','2013-01-01','2013-06-30', timezone='CST')
FiseraTobs <- WISKItoObs(FiseraTvalues, timezone='Etc/GMT-7')
}
\seealso{
\code{\link{getWISKIvalues}}
}
\author{
Kevin Shook
}
