% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWISKIvalues.R
\name{getWISKIvalues}
\alias{getWISKIvalues}
\title{Gets values of a WISKI time series}
\usage{
getWISKIvalues(timeSeries = "", startDate = "1900-01-01", endDate = "",
  timezone = "", site.url = "http://giws.usask.ca:8080/")
}
\arguments{
\item{timeSeries}{Required. Character string containing the WISKI time series ID number, which is returned by \code{findWISKItimeseries}. Cannot contain wild card characters.}

\item{startDate}{Optional. Character vector of the starting date of data being queried. Must be in the form \code{'yyyy-mm-dd'}. The default value is \code{'1900-01-01'}.}

\item{endDate}{Optional. Character vector of the ending date of data being queried. Must be in the form \code{'yyyy-mm-dd'}. The default value is today's date.}

\item{timezone}{Required. Time zone for the data. Can be any time zone string sed by \code{Java} such as \option{GMT-5}, or \option{MST}.}

\item{site.url}{Optional. A character string containing the url of the WISKI web server. Defaults to the Global Institute for Water Security (GIWS) server: \code{'http://giws.usask.ca:8080/'}.  As this package is intended for use by the GIWS hydrological community, it is usually unnecessary to specify the web server.}
}
\value{
If unsuccessful, returns \code{FALSE}. If successful, returns a dataframe with three variables: \item{time}{R time value} \item{variable name}{time series values} \item{QualityCode}{time series quality codes}
}
\description{
Returns times, values, and quality codes for a specified interval of a specified time series.
}
\examples{
# get values for Fisera Ridge air temperatures (Original time series) 
FiseraTvalues <- getWISKIvalues('9296042','2013-01-01','2013-06-30', timezone='CST')
summary(FiseraTvalues)
}
\seealso{
\code{\link{getWISKImetadata}} \code{\link{findWISKItimeseries}}
}
\author{
Kevin Shook
}
